==> croisement_data/Step2_answer.Rmd <==
---
title: "Step2b_trueanswer"
author: "AIGOIN Emilie"
date: "2025-04-02"
output: html_document
---

```{r, warning = FALSE}

# Charger les packages

library(jsonlite)
library(data.table)

```

```{r}

# Mise en place du dossier de travail

# Définir le chemin vers le dossier de travail : !!!! à changer !!!!
chemin <- "/Users/emilieaigoin/Desktop/Test-ordinateur/multiple_data"

# Changer le répertoire actuel vers ce dossier
setwd(chemin)

# Afficher le répertoire de travail pour confirmation
cat("Répertoire de travail actuel:", getwd(), "\n")

```

## Chargement des données

```{r}

# Mise en place du fichier des experts
expert <- fread(file.path(chemin, "ground_truth.txt"), header = FALSE)

# Enlever la première ligne (car en trop)
expert <- expert[-1,]

# Ajouter le numéro de ligne comme colonne
expert[, ligne := .I]

# Filtrer les lignes où la valeur n'est pas -1
expert_filtre <- expert[V1 != -1]

# Afficher les indices des lignes qui ne sont pas -1
indices_valides <- expert_filtre$ligne

```

```{r}

# Lecture du fichier tasks
tasks <- fromJSON(file.path(chemin, "tasks.json"), simplifyVector = TRUE)

# Convertir en data.table (traitement plus rapide)
tasks_dt <- data.table(
  id = names(tasks),
  numero = as.integer(unname(tasks))
)

```

## Chercher les correspondances

```{r}

# Si vous préférez une jointure complète avec toutes les données de ground_truth_filtre:
resultats <- expert_filtre[tasks_dt, on = .(ligne = numero), nomatch = NULL]

```

## Convertir les résultats en .json

```{r}

# Créer la structure JSON souhaitée
json_output <- list()

# Pour chaque ligne du dataframe
for (i in 1:nrow(resultats)) {
  
  # Extraire les valeurs de chaque colonne pour les renommer 
  identifiant <- as.character(resultats[i, id])
  valeur_obs <- as.character(resultats[i, ligne])
  valeur_id <- as.integer(resultats[i, V1])
  
  # Ajouter à notre structure JSON dans le format souhaité
  json_output[[identifiant]] <- list(
    obs_SWE = valeur_obs,
    id_SWE = valeur_id
  )
}

# Convertir en format JSON
json_string <- toJSON(json_output, pretty = TRUE, auto_unbox = TRUE)

# Enregistrer les résultats
writeLines(json_string, "3_true_classes.json")

```
